gitk: Add options --select-file and --select-line
These can be used in combination with --select-commit to jump to a given
line in a patch on startup. (They don't have any effect when
--select-commit is not used.)

This is useful for the "Show History Context" command in Git Gui's blame
window, which currently only jumps to the right commit in gitk, but
doesn't select the line that the context menu was opened on.

Also, these options allow for powerful editor integration; they make it
possible to jump into gitk right from a text editor. For example, here's
a small ruby script that takes a file path and a line number and opens
gitk with the commit selected that last modified that line. This can
easily be mapped to a key in vim or other editors.

    #!/usr/bin/env ruby

    if ARGV.length != 2
        puts "Usage: #{$0} <file> <line>"
        exit 1
    end

    file, line = ARGV
    blame_output = `git blame -p -L#{line},+1 "#{file}"`
    exit 1 if $?.exitstatus != 0

    blame_output_lines = blame_output.split("\n")
    commit, line = blame_output_lines[0].split

    file = blame_output_lines.grep(/^filename /)[0][9..-1]
    date = blame_output_lines.grep(/^committer-time /)[0][15..-1]
    two_weeks_later = date.to_i + 60 * 60 * 24 * 7 * 2

    system "gitk --before='#{two_weeks_later}' \
                 --select-commit=#{commit} \
                 --select-file='#{file}' \
                 --select-line=#{line} &"