git-gui: guard against "Application error" when using arrow keys
In rare cases, pressing up-arrow or down-arrow in the list of unstaged or staged
files would result in an "Application Error" popup with the text
'expected number but got ""'. One such situation was: if there's one modified
(unstaged) file and one untracked file, and you click on the modified file and
press Command-T to stage it, then the untracked file isn't selected. If you then
press up-arrow or down-arrow, the code that's changed here would try to find a
selected file, assuming there must be one. Make this more robust by returning 0
in this case, so that pressing up or down starts from the top.

Signed-off-by: Stefan Haller <stefan@haller-berlin.de>