git-gui: stage or unstage files with double-click or Return
When observing new users of git gui, I often see that they have trouble finding
out how to stage or unstage files. It is not very intuitive that you have to
click the small icon next to the file name for that. Many users try
double-clicking a file to stage it, so support that.

Pressing Return to stage files is something that I simply find convenient
myself.

Signed-off-by: Stefan Haller <stefan@haller-berlin.de>